#357 move common code out of transformation scripts and into new file
